Crate rspirv2_types

rspirv2-types

Defines types and traits to describe SPIR-V Instructions and Operands. Part of rspirv2.

The crate doesn’t include the core grammar or any of the extended instruction sets defined by the SPIR-V grammar json, it merely defines the types to represent them. The main crate rspirv2 reexports this crate and includes auto generated definitions of the core grammar, with some select Vulkan-specific extended instruction sets.

Modules

binary

dis

Module for Disassembly

inst

meta

Rust structs containing metadata information about SPIR-V instructions, directly deserialized from the SPIR-V grammar JSON.

module

operand

slice

vec

Structs

Word

A 32bit SPIR-V Word

Functions

cast_words_to_ne_bytes

Convert a slice of Words to a slice of u8s in native endian.