rspirv2_types

Module binary

Structs

DecodeError

FnWriter

IdResultAlloc

A sharable atomic IdResultAlloc, just Clone to share it

InstOffset

An offset to some instruction, a wrapper around usize.

InstReader

Reader for a single instruction

OperandReader

Read operand Words from an instruction

WordCounter

Counts the amount of Words that were written, discarding the words itself

WordSliceWriter

Enums

DecodeErrorKind

EncodeError

Constants

ID_RESULT_MAX

A reasonable upper bounds that ensures subsequent allocations by potentially multiple threads all fail

Traits

WordWriter

A WordWriter allows you to Self::write Words into it and is usually backed by a Vec.