rspirv2_types

Module dis

Module for Disassembly

Structs

DisContext

Context object for disassembly generation

DisInstSlice

A sequence of words that has been pre-processed and may be Displayed.

DisOptions

Options for disassembly

SeparatorJoiner

Utility struct for inserting separators between variants. Use it like format!({sep}{value}), will skip the separator on first write.

StrCharIter

Like a str::chars Iterator, but yields &str of the original &str instead

Enums

IdName

IdNaming

Defines where names of types should be derived from.

PrimitiveType

An enum for representing primitive types

ResolvedIdName

TypeNaming

Defines how to derive names for types

Constants

STR_15_SPACES

&str of 15 spaces, the default padding in spirv-tools

Traits

InstSetDisCtx

An instruction set that provides additional context information for disassembly.

IntoDisContext

Either DisContext or DisOptions that is turned into one via DisContext::no_context. Actual context needs to be gathered yourself, see warning in DisContext::no_context.

Functions

escape_cow

escape chars individually with another char

escape_cow_str

escape chars individually with a &str that may contain multiple chars

escape_id_name