rspirv2_types::dis

Struct DisContext

pub struct DisContext {
    pub id_to_int_width: FxHashMap<IdResult, u32>,
    pub id_type_to_primitive_type: FxHashMap<IdResult, PrimitiveType>,
    pub id_to_const_fmt: FxHashMap<IdResult, ConstFmt>,
    pub id_to_name: FxHashMap<IdResult, IdName>,
    /* private fields */
}

Context object for disassembly generation

Next to DisOptions, contains lookup tables to aid in disassembly generation. Initializing these tables is required to generate correct disassembly that spirv-as can read, otherwise, we’ll generate the best possible disassembly we can. Since instructions are not available within the rspirv2-types crate, we can’t actually implement the context information retrieval here. Instead, it needs to be implemented for each instruction set with the InstSetDisCtx trait, which may delegate to our default implementation in rspirv2::dis::create_dis_context_core.

Fields
id_to_int_width: FxHashMap<IdResult, u32>

Maps an IdResult of an arbitrary instruction to its integer width, if it is an integer. To decode an OpSwitch, one needs to know from context whether the selector is 32 or 64bit, as it determines whether the label constants consume one or two words.

id_type_to_primitive_type: FxHashMap<IdResult, PrimitiveType>

Maps an IdResult of a type declaration to a PrimitiveType. Primarily for OpSwitch, but such a lookup may be useful for other instructions as well.

id_to_const_fmt: FxHashMap<IdResult, ConstFmt>

Maps an IdResult of a type declaration to a ConstFmt to tell OpConstant instructions how to format the untyped constant value

id_to_name: FxHashMap<IdResult, IdName>

Maps an IdResult to the “highest priority” IdName

Implementations

impl DisContext

pub fn no_context(opt: DisOptions) -> Self

Creates a new DisContext without having scanned the module for the required extra information.

WARNING: You need to Self::add_context the instructions you want to decode before disassembling, as the disassembly generated may be invalid without the required context. For example, OpConstant needs to query the type of constant their value is to format it correctly as a float or an int.

pub fn add_context<ISA: InstSetDisCtx>(&mut self, slice: &InstSlice<ISA>)

Scan the supplied InstSlice for useful context

pub fn add_context_raw<ISA: InstSetDisCtx>(&mut self, slice: &RawInstSlice)

Scan the supplied InstSlice for useful context, skip over any DecodeErrors that may arise

pub fn add_id_to_name(&mut self, id: IdResult, name: IdName)

Add a mapping from an IdResult to an IdName

pub fn id_to_name(&self, id: IdResult) -> ResolvedIdName<'_>