rspirv2_types

Module operand

Structs

FixedLenComposer

Compose the OperandEncoding::FIXED_LEN from multiple maybe fixed len Operands

FormattedValue

IdMemorySemantics

A IdMemorySemantics is a reference to a IdResult of another operation.

IdRef

A IdRef is a reference to a IdResult of another operation.

IdResult

A SPIR-V “Result ID”.

IdResultType

A IdResultType is a reference to a IdResult of another operation.

IdResultWriter

IdScope

A IdScope is a reference to a IdResult of another operation.

LiteralConst

A LiteralContextDependentNumber, or LiteralConst for short since it’s only used by OpConstant (and OpSpecConstant) instructions.

LiteralExtInstInteger

The Instruction ID from an extended instruction set, backed by a 32bit integer.

LiteralFloat

A float literal as defined by SPIR-V spec: simply an f32. It’s represented as a u32 to ensure the bit pattern isn’t changing.

LiteralInteger

An integer literal as defined by SPIR-V spec: a 32bit integer.

LiteralSpecConstantOpInteger

The Instruction ID for an OpSpecConstantOp, backed by a 32bit integer.

LiteralString

A SPIR-V String literal. Defined as a sequence of UTF-8, so we can just use an ordinary String.

OperandDis

OperandDisContext

Context object for disassembly generation of operands within an instruction

ParameterizedBitmask

Enums

ConstFmt

How the untyped constant value of a LiteralConst should be formatted

LiteralStringEscape

Describes how to escape string sequences

Constants

ID_RESULT_COLOR

OPERAND_KIND_ID_MEMORY_SEMANTICS

OPERAND_KIND_ID_REF

OPERAND_KIND_ID_RESULT

OPERAND_KIND_ID_RESULT_TYPE

OPERAND_KIND_ID_SCOPE

OPERAND_KIND_LITERAL_CONTEXT_DEPENDENT_NUMBER

OPERAND_KIND_LITERAL_EXT_INST_INTEGER

OPERAND_KIND_LITERAL_FLOAT

OPERAND_KIND_LITERAL_INTEGER

OPERAND_KIND_LITERAL_SPEC_CONSTANT_OP_INTEGER

OPERAND_KIND_LITERAL_STRING

PARAMETERIZED_BITMASK_REQUIRES_FIXED_LEN

Traits

Operand

A SPIR-V operand. The associated const Self::KIND links to it’s OperandKind.

OperandEncoding

Something that can be decoded from or encoded to SPIR-V, not necessarily a full Operand.

ParameterizedBitmaskBits

Type Aliases

LiteralContextDependentNumber

ZeroOrMore

ZeroOrOne