rspirv2_types::operand

Struct LiteralConst

pub struct LiteralConst(/* private fields */);

A LiteralContextDependentNumber, or LiteralConst for short since it’s only used by OpConstant (and OpSpecConstant) instructions.

A number of content dependent length, as defined by SPIR-V spec. May represent any number of Words, depending on the type of the constant.

Parsing Assumption

We assume that LiteralConst is always the last Operand in an Instruction and never has a quantity of Quantifier::ZeroOrMore.

The current spec satisfies this requirement. OpConstant / OpSpecConstant are the only instructions to consume a LiteralConst as the last operant and expect exactly one operant.

This greatly simplifies parsing, as we can simply assume the remaining words of this instruction all contribute to the constant operand. The “correct” way to handle this would be to parse the “result type id”, resolve its type from previously parsed instructions and compute its size from the type layout specification. This would make the implementation vastly more complex and likely have a negative impact on decoding performance, which is why we decided to make this assumption about SPIR-V grammars.

A LiteralConst can only be decoded with OperandEncoding::decode_last. Decoding an LiteralInteger not as the last operand, aka. calling OperandEncoding::decode, will always return an Error.

Implementations

impl LiteralConst

pub fn words(&self) -> usize

impl LiteralConst

pub fn as_u64(&self) -> Result<u64, DecodeError>

pub fn as_u32(&self) -> Result<u32, DecodeError>

pub fn as_u16(&self) -> Result<u16, DecodeError>

pub fn as_u8(&self) -> Result<u8, DecodeError>

impl LiteralConst

pub fn as_i8(&self) -> Result<i8, DecodeError>

pub fn as_i16(&self) -> Result<i16, DecodeError>

pub fn as_i32(&self) -> Result<i32, DecodeError>

pub fn as_i64(&self) -> Result<i64, DecodeError>

pub fn as_f32(&self) -> Result<f32, DecodeError>

pub fn as_f64(&self) -> Result<f64, DecodeError>

impl LiteralConst

pub fn fmt_value<'a>( &'a self, ctx: &'a OperandDisContext<'a>, ) -> FormattedValue<'a>